Correction of some mistakes

+++ src/main/population/create-schema.sql
@@ -124,7 +124,7 @@
         `investing_statement` varchar(255),
         `name` varchar(255),
         `sector` varchar(255),
-        `stars` integer,
+        `stars` double precision,
         primary key (`id`)
     ) engine=InnoDB;
